+++ CHANGES.rst
@@ -1,7 +1,14 @@
 List of releases of Lasco
 =========================
 
-Lasco 0.1.0 (to be released)
-----------------------------
+unreleased
+----------
+
+- fixed completion of ``help`` command in the command-line
+  interface. All commands appeared twice.
+
+
+Lasco 0.1.0 (2012-06-19)
+------------------------
 
 First public release.